Infer returns an InferenceDataSet object - Also moved infer to the new class-based verb infrastructure
